Add cache support.
Right now for every single query we go to zk and fetch all children
data then do an in-mem filtering. When we have thousands of children
this becomes a very expensive operation. This PR adds an optional
feature of caching fetched result from zk. The way this cache is
implemented is a compromise between consistency and complexity.

Ideally to make cache 100% consistent we should subcribe to every
single change to children's data. However, in reality this is
inpractical because it's possible there could be children updates
all the time due to periodical jobs.

So this PR uses a daemon that refreshes cache every N (tunable)
seconds. In addition, we subscript to all children joining/leaving
events. This is less frequent because normally no one would constantly
add/remove machines. So whenever a join/leave event happens, we immediately
refresh the cache. This way we guarantee that whenever we add/remove
machines, cache will always have the right set of machines.